third_party/blink/web_tests/wpt_internal/webmcp/execute_tool_submit_from_js.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<link rel="author" href="mailto:masonf@chromium.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/webmcp-helpers.js"></script>

<form toolname="submit_test" tooldescription="Submit test form" target="hidden_iframe" action="about:blank" toolautosubmit>
</form>
<iframe name="hidden_iframe" style="display:none"></iframe>

<script>
promise_test(async t => {
  await waitForTool('submit_test');

  let form = document.querySelector('form');
  let submitFired = false;

  form.addEventListener('submit', (e) => {
    submitFired = true;
    e.preventDefault();
    form.submit();
  });

  // Call executeTool. If it throws, the test fails automatically.
  // The old bug was that this would reject with a "programming error" string.
  await navigator.modelContextTesting.executeTool('submit_test', '{}');

  assert_true(submitFired, 'submit event should have fired');
}, 'executeTool resolves successfully when submit handler calls form.submit()');
</script>
